+++ tests/module_disambiguation/errors/source/all_substitutions_equal.rs
@@ -1,0 +1,10 @@
+use duplicate::*;
+
+#[duplicate_item(
+		sub_typ 		sub_fn;
+		[ SomeType ] 	[ some_fn ];
+		[ SomeType  ]	[ some_fn ];
+)]//duplicate_end
+mod __
+{}
+//item_end